corrected bug(sp speed to 0, changed state machien order, disabled timers logic)

+++ motor/mc_interface.c
@@ -3038,6 +3038,6 @@ float mc_interface_get_bike_set_rpm(void) {
 	return mcpwm_foc_get_bike_set_rpm();
 }
 float mc_interface_get_bike_control_mode(void){
-	return mcpwm_foc_get_bike_control_mode();
+	return (double)mcpwm_foc_get_bike_control_mode();
 
 }
